Added ids for keywords and also messages
The start_keyword and end_keyword methods does not provide ids. This
invalidates the model where you can have a stateless server that
receive information about the execution. Issue #2115 in robotframework
github page have more information about it.

This is adding a id generator to the listener interface that allows
us to keep working on the robot-dashboard project.